requirements.txt
asgiref==3.5.2
atapt==0.4
certifi==2022.5.18.1
charset-normalizer==3.2.0
distlib==0.3.4
Django==4.0.4
django-ckeditor==6.5.1
django-debug-toolbar==3.4.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.7.0
galois==0.3.6
humanfriendly==10.0
idna==3.4
llvmlite==0.41.0
Markdown==3.4.1
mpmath==1.3.0
numba==0.58.0
numpy==1.22.0
pandas==2.0.1
Pillow==9.3.0
pipenv==2022.5.2
platformdirs==2.5.2
psycopg2==2.9.5
pyreadline3==3.4.1
pySMART==1.2.4
python-dateutil==2.8.2
pytz==2022.7
pywin32==306
requests==2.31.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.2
sympy==1.12
typing_extensions==4.8.0
tzdata==2022.1
unireedsolomon==1.0.6
urllib3==2.0.4
virtualenv==20.14.1
virtualenv-clone==0.5.7
wget==3.2
